+++ test/Telegram.Bot.Tests.Integ/Admin Bot/ChannelAdminBotTests.cs
@@ -35,7 +35,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSetChatTitle);
 
             await BotClient.SetChatTitleAsync(
-                chatId: _classFixture.ChatId,
+                chatId: _classFixture.Chat.Id,
                 title: _classFixture.ChatTitle
             );
         }
@@ -51,7 +51,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldSetChatDescription);
 
             await BotClient.SetChatDescriptionAsync(
-                chatId: _classFixture.ChatId,
+                chatId: _classFixture.Chat.Id,
                 description: "Test Chat Description"
             );
         }
@@ -62,7 +62,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldDeleteChatDescription);
 
-            await BotClient.SetChatDescriptionAsync(_classFixture.ChatId);
+            await BotClient.SetChatDescriptionAsync(_classFixture.Chat.Id);
         }
 
         #endregion
@@ -75,10 +75,10 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldPinMessage);
 
-            Message msg = await BotClient.SendTextMessageAsync(_classFixture.ChatId, "Description to pin");
+            Message msg = await BotClient.SendTextMessageAsync(_classFixture.Chat.Id, "Description to pin");
 
             await BotClient.PinChatMessageAsync(
-                chatId: _classFixture.ChatId,
+                chatId: _classFixture.Chat.Id,
                 messageId: msg.MessageId
             );
 
@@ -93,7 +93,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
 
             Message pinnedMsg = _classFixture.PinnedMessage;
 
-            Chat chat = await BotClient.GetChatAsync(_classFixture.ChatId);
+            Chat chat = await BotClient.GetChatAsync(_classFixture.Chat.Id);
 
             Assert.True(JToken.DeepEquals(
                 JToken.FromObject(pinnedMsg), JToken.FromObject(chat.PinnedMessage)
@@ -106,7 +106,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldUnpinMessage);
 
-            await BotClient.UnpinChatMessageAsync(_classFixture.ChatId);
+            await BotClient.UnpinChatMessageAsync(_classFixture.Chat.Id);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldGetChatWithNoPinnedMessage)]
@@ -115,7 +115,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldGetChatWithNoPinnedMessage);
 
-            Chat chat = await BotClient.GetChatAsync(_classFixture.ChatId);
+            Chat chat = await BotClient.GetChatAsync(_classFixture.Chat.Id);
 
             Assert.Null(chat.PinnedMessage);
         }
@@ -133,7 +133,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             using (Stream stream = System.IO.File.OpenRead(Constants.FileNames.Photos.Logo))
             {
                 await BotClient.SetChatPhotoAsync(
-                    chatId: _classFixture.ChatId,
+                    chatId: _classFixture.Chat.Id,
                     photo: stream
                 );
             }
@@ -145,7 +145,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldGetChatPhoto);
 
-            Chat chat = await BotClient.GetChatAsync(_classFixture.ChatId);
+            Chat chat = await BotClient.GetChatAsync(_classFixture.Chat.Id);
 
             Assert.NotEmpty(chat.Photo.BigFileId);
             Assert.NotEmpty(chat.Photo.SmallFileId);
@@ -157,7 +157,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldDeleteChatPhoto);
 
-            await BotClient.DeleteChatPhotoAsync(_classFixture.ChatId);
+            await BotClient.DeleteChatPhotoAsync(_classFixture.Chat.Id);
         }
 
         [OrderedFact(DisplayName = FactTitles.ShouldThrowOnDeletingChatDeletedPhoto)]
@@ -167,10 +167,42 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldThrowOnDeletingChatDeletedPhoto);
 
             Exception e = await Assert.ThrowsAnyAsync<Exception>(() =>
-                BotClient.DeleteChatPhotoAsync(_classFixture.ChatId));
+                BotClient.DeleteChatPhotoAsync(_classFixture.Chat.Id));
 
             Assert.IsType<ApiRequestException>(e);
             Assert.Equal("Bad Request: CHAT_NOT_MODIFIED", e.Message);
+        }
+
+        /// <summary>
+        /// If chat had a photo before, reset the photo back.
+        /// </summary>
+        [OrderedFact(DisplayName = FactTitles.ShouldResetOldChatPhoto)]
+        [Trait(Constants.MethodTraitName, Constants.TelegramBotApiMethods.SetChatPhoto)]
+        public async Task Should_Reset_Old_Chat_Photo_If_Existed()
+        {
+            await _fixture.SendTestCaseNotificationAsync(FactTitles.ShouldResetOldChatPhoto);
+
+            // "Chat.Photo" might be null if there is no photo currently set
+            string previousChatPhotoId = _classFixture.Chat.Photo?.BigFileId;
+            if (previousChatPhotoId == default)
+            {
+                // chat didn't have a photo
+                return;
+            }
+
+            using (Stream photoStream = new MemoryStream())
+            {
+                // pass photo's file_id, prepare file for download, and download the file into memroy
+                await BotClient.GetInfoAndDownloadFileAsync(previousChatPhotoId, photoStream);
+
+                // need to set position of memory stream back to its start so next method reads photo stream from the beginning
+                photoStream.Position = 0;
+
+                await BotClient.SetChatPhotoAsync(
+                    chatId: _classFixture.Chat.Id,
+                    photo: photoStream
+                );
+            }
         }
 
         #endregion
@@ -186,7 +218,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             const string setName = "EvilMinds";
 
             ApiRequestException exception = await Assert.ThrowsAnyAsync<ApiRequestException>(() =>
-                _fixture.BotClient.SetChatStickerSetAsync(_classFixture.ChatId, setName)
+                _fixture.BotClient.SetChatStickerSetAsync(_classFixture.Chat.Id, setName)
             );
 
             // ToDo: Create exception type
@@ -221,6 +253,8 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
             public const string ShouldThrowOnDeletingChatDeletedPhoto =
                 "Should throw exception in deleting chat photo with no photo currently set";
 
+            public const string ShouldResetOldChatPhoto = "Should reset the same old chat photo if existed";
+
             public const string ShouldThrowOnSetChannelStickerSet =
                 "Should throw exception when trying to set sticker set for a channel";
         }
@@ -229,7 +263,7 @@ namespace Telegram.Bot.Tests.Integ.Admin_Bot
         {
             public Fixture(TestsFixture fixture)
             {
-                ChatId = new ChannelChatFixture(fixture, Constants.TestCollections.ChannelAdminBots).ChannelChatId;
+                Chat = new ChannelChatFixture(fixture, Constants.TestCollections.ChannelAdminBots).ChannelChat;
             }
         }
     }