+++ test/Telegram.Bot.Tests.Integ/Framework/UpdateReceiver.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
@@ -10,14 +10,14 @@ namespace Telegram.Bot.Tests.Integ.Framework
 {
     public class UpdateReceiver
     {
+        public ICollection<string> AllowedUsernames { get; }
+
         private readonly ITelegramBotClient _botClient;
 
-        private readonly string[] _allowedUsernames;
-
-        public UpdateReceiver(ITelegramBotClient botClient, params string[] allowedUsernames)
+        public UpdateReceiver(ITelegramBotClient botClient, IEnumerable<string> allowedUsernames)
         {
             _botClient = botClient;
-            _allowedUsernames = allowedUsernames;
+            AllowedUsernames = new List<string>(allowedUsernames);
         }
 
         public async Task DiscardNewUpdatesAsync(CancellationToken cancellationToken = default)
@@ -33,17 +33,13 @@ namespace Telegram.Bot.Tests.Integ.Framework
             while (!cancellationToken.IsCancellationRequested)
             {
                 var updates = await _botClient.GetUpdatesAsync(offset,
-                    allowedUpdates: new UpdateType[0],
+                    allowedUpdates: Array.Empty<UpdateType>(),
                     cancellationToken: cancellationToken);
 
-                if (updates.Any())
-                {
-                    offset = updates.Last().Id + 1;
-                }
-                else
-                {
+                if (updates.Length == 0)
                     break;
-                }
+
+                offset = updates.Last().Id + 1;
             }
         }
 
@@ -54,39 +50,26 @@ namespace Telegram.Bot.Tests.Integ.Framework
             params UpdateType[] updateTypes)
         {
             if (cancellationToken == default)
-            {
-                var source = new CancellationTokenSource(TimeSpan.FromMinutes(2));
-                cancellationToken = source.Token;
-            }
+                cancellationToken = new CancellationTokenSource(TimeSpan.FromMinutes(2)).Token;
 
-            Update[] matchingUpdates = default;
+            Update[] matchingUpdates = Array.Empty<Update>();
 
             while (!cancellationToken.IsCancellationRequested)
             {
                 IEnumerable<Update> updates = await GetOnlyAllowedUpdatesAsync(offset, cancellationToken, updateTypes);
-
-                if (predicate == null)
-                {
-                    updates = updates.Where(u => updateTypes.Contains(u.Type));
-                }
-                else
-                {
-                    updates = updates
-                        .Where(u => updateTypes.Contains(u.Type))
-                        .Where(predicate);
-                }
+                updates = updates
+                    .Where(u =>
+                        updateTypes.Contains(u.Type) &&
+                        (predicate == null || predicate(u))
+                    );
 
                 matchingUpdates = updates.ToArray();
 
                 if (updates.Any())
-                {
                     break;
-                }
-                else
-                {
-                    offset = updates.LastOrDefault()?.Id + 1 ?? 0;
-                    await Task.Delay(1_500, cancellationToken);
-                }
+
+                offset = updates.LastOrDefault()?.Id + 1 ?? 0;
+                await Task.Delay(1_500, cancellationToken);
             }
 
             return matchingUpdates;
@@ -98,28 +81,16 @@ namespace Telegram.Bot.Tests.Integ.Framework
             bool discardNewUpdates = true,
             CancellationToken cancellationToken = default)
         {
-            Func<Update, bool> predicate = null;
-            if (messageId != default && data != default)
-            {
-                predicate = u => u.CallbackQuery.Message.MessageId == messageId &&
-                                 u.CallbackQuery.Data == data;
-            }
-            else
-            {
-                if (messageId != default)
-                    predicate = u => u.CallbackQuery.Message.MessageId == messageId;
-                else if (data != default)
-                    predicate = u => u.CallbackQuery.Data == data;
-            }
+            Func<Update, bool> predicate = u =>
+                (messageId == default || u.CallbackQuery.Message.MessageId == messageId) &&
+                (data == default || u.CallbackQuery.Data == data);
 
             var updates = await GetUpdatesAsync(predicate,
                 cancellationToken: cancellationToken,
                 updateTypes: UpdateType.CallbackQuery);
 
             if (discardNewUpdates)
-            {
                 await DiscardNewUpdatesAsync(cancellationToken);
-            }
 
             var update = updates.First();
             return update;
@@ -133,9 +104,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
                 updateTypes: UpdateType.InlineQuery);
 
             if (discardNewUpdates)
-            {
                 await DiscardNewUpdatesAsync(cancellationToken);
-            }
 
             var update = updates.First();
             return update;
@@ -162,7 +131,7 @@ namespace Telegram.Bot.Tests.Integ.Framework
                 var updates = await GetUpdatesAsync(
                     u => u.Message?.Type == messageType || u.ChosenInlineResult != null,
                     cancellationToken: cancellationToken,
-                    updateTypes: new[] { UpdateType.Message, UpdateType.ChosenInlineResult }
+                    updateTypes: new[] {UpdateType.Message, UpdateType.ChosenInlineResult}
                 );
 
                 messageUpdate = updates.SingleOrDefault(u => u.Message?.Type == messageType);
@@ -175,62 +144,36 @@ namespace Telegram.Bot.Tests.Integ.Framework
         private async Task<Update[]> GetOnlyAllowedUpdatesAsync(
             int offset, CancellationToken cancellationToken, params UpdateType[] types)
         {
-            var updates = await _botClient.GetUpdatesAsync(offset,
+            var updates = await _botClient.GetUpdatesAsync(
+                offset,
                 allowedUpdates: types,
                 cancellationToken: cancellationToken);
 
             var allowedUpdates = updates.Where(IsAllowed).ToArray();
-
             return allowedUpdates;
         }
 
         private bool IsAllowed(Update update)
         {
-            if (_allowedUsernames is null || _allowedUsernames.All(string.IsNullOrWhiteSpace))
-            {
+            if (AllowedUsernames is null || AllowedUsernames.All(string.IsNullOrWhiteSpace))
                 return true;
-            }
-
-            bool isAllowed;
 
             switch (update.Type)
             {
                 case UpdateType.Message:
-                    isAllowed = _allowedUsernames
-                        .Contains(update.Message.From.Username, StringComparer.OrdinalIgnoreCase);
-                    break;
                 case UpdateType.InlineQuery:
-                    isAllowed = _allowedUsernames
-                        .Contains(update.InlineQuery.From.Username, StringComparer.OrdinalIgnoreCase);
-                    break;
                 case UpdateType.CallbackQuery:
-                    isAllowed = _allowedUsernames
-                        .Contains(update.CallbackQuery.From.Username, StringComparer.OrdinalIgnoreCase);
-                    break;
                 case UpdateType.PreCheckoutQuery:
-                    isAllowed = _allowedUsernames
-                        .Contains(update.PreCheckoutQuery.From.Username, StringComparer.OrdinalIgnoreCase);
-                    break;
                 case UpdateType.ShippingQuery:
-                    isAllowed = _allowedUsernames
-                        .Contains(update.ShippingQuery.From.Username, StringComparer.OrdinalIgnoreCase);
-                    break;
                 case UpdateType.ChosenInlineResult:
-                    isAllowed = _allowedUsernames
-                        .Contains(update.ChosenInlineResult.From.Username, StringComparer.OrdinalIgnoreCase);
-                    break;
+                    return AllowedUsernames.Contains(update.GetUser().Username, StringComparer.OrdinalIgnoreCase);
                 case UpdateType.EditedMessage:
                 case UpdateType.ChannelPost:
                 case UpdateType.EditedChannelPost:
-                    isAllowed = false;
-                    break;
-                case UpdateType.Unknown:
-                    throw new ArgumentException("Unknown update found!");
+                    return false;
                 default:
                     throw new ArgumentOutOfRangeException();
             }
-
-            return isAllowed;
         }
     }
 }